token stored in db

+++ src/main/java/whilter/ai/ads_manager/model/AdSetRequest.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class AdSetRequest {
     private String name;
 //    private String campaignId;
-    private float dailyBudget;
+    private int dailyBudget;
     private String billingEvent;
     private String optimizationGoal;
     private String status;
